# Deduplicate transducer contracts