… private/transducer-base.rkt