Improves #351

Because there are contracts on the creator and accessors for the
transducer callbacks each applies the same contract on the callbacks.
This change weakens the contract on the accessors, and should not change
any contractual obligations provided transducers are created with
`make-transducer`.

This provides modest performance improvements:

- [Vector to List - 1000000 elements](https://gist.github.com/samdphillips/92dd56e1fc78884230dbc59aabf6f1c0#file-vector-list-rkt)
  - 32s -> 22.4s
- [1000000 Range to Void](https://gist.github.com/samdphillips/92dd56e1fc78884230dbc59aabf6f1c0#file-identity-rkt)
  - 30s -> 19.8s
- [1000000 Range pass through 5 `mapping` to Void](https://gist.github.com/samdphillips/92dd56e1fc78884230dbc59aabf6f1c0#file-map5-rkt)
  - 283s -> 206s