@jackfirth 2019-12-18
This is great! Merging. Also, I wrote some more [benchmarks](https://gist.github.com/jackfirth/9596dd48536344b4b94c1938a36de4c9) for transducers. I haven't measured how this PR affects them though.